Move permissions module into timed module
This way it is accessible by all apps.
